move various symbols to cli/node/mod.rs
