+++ cli/compat/mod.rs
@@ -3,6 +3,8 @@
 mod esm_resolver;
 
 use crate::file_fetcher::FileFetcher;
+use crate::node::NODE_COMPAT_URL;
+use crate::node::SUPPORTED_MODULES;
 use deno_ast::MediaType;
 use deno_core::error::AnyError;
 use deno_core::located_script_name;
@@ -13,63 +15,6 @@ use once_cell::sync::Lazy;
 
 pub use esm_resolver::check_if_should_use_esm_loader;
 pub use esm_resolver::NodeEsmResolver;
-
-// WARNING: Ensure this is the only deno_std version reference as this
-// is automatically updated by the version bump workflow.
-pub(crate) static STD_URL_STR: &str = "https://deno.land/std@0.153.0/";
-
-static SUPPORTED_MODULES: &[&str] = &[
-  "assert",
-  "assert/strict",
-  "async_hooks",
-  "buffer",
-  "child_process",
-  "cluster",
-  "console",
-  "constants",
-  "crypto",
-  "dgram",
-  "dns",
-  "domain",
-  "events",
-  "fs",
-  "fs/promises",
-  "http",
-  "https",
-  "module",
-  "net",
-  "os",
-  "path",
-  "path/posix",
-  "path/win32",
-  "perf_hooks",
-  "process",
-  "querystring",
-  "readline",
-  "stream",
-  "stream/promises",
-  "stream/web",
-  "string_decoder",
-  "sys",
-  "timers",
-  "timers/promises",
-  "tls",
-  "tty",
-  "url",
-  "util",
-  "util/types",
-  "v8",
-  "vm",
-  "worker_threads",
-  "zlib",
-];
-
-static NODE_COMPAT_URL: Lazy<String> = Lazy::new(|| {
-  std::env::var("DENO_NODE_COMPAT_URL")
-    .map(String::into)
-    .ok()
-    .unwrap_or_else(|| STD_URL_STR.to_string())
-});
 
 static GLOBAL_URL_STR: Lazy<String> =
   Lazy::new(|| format!("{}node/global.ts", NODE_COMPAT_URL.as_str()));
@@ -95,20 +40,6 @@ static COMPAT_IMPORT_URL: Lazy<Url> =
 /// Provide imports into a module graph when the compat flag is true.
 pub fn get_node_imports() -> Vec<(Url, Vec<String>)> {
   vec![(COMPAT_IMPORT_URL.clone(), vec![GLOBAL_URL_STR.clone()])]
-}
-
-pub fn try_resolve_builtin_module(specifier: &str) -> Option<Url> {
-  if SUPPORTED_MODULES.contains(&specifier) {
-    let ext = match specifier {
-      "stream/promises" => "mjs",
-      _ => "ts",
-    };
-    let module_url =
-      format!("{}node/{}.{}", NODE_COMPAT_URL.as_str(), specifier, ext);
-    Some(Url::parse(&module_url).unwrap())
-  } else {
-    None
-  }
 }
 
 pub fn load_cjs_module(
