+++ cli/tools/init/mod.rs
@@ -1,8 +1,9 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
 use crate::args::InitFlags;
-use crate::compat;
-use deno_core::{anyhow::Context, error::AnyError};
+use crate::node;
+use deno_core::anyhow::Context;
+use deno_core::error::AnyError;
 use std::io::Write;
 use std::path::Path;
 
@@ -35,7 +36,7 @@ pub async fn init_project(init_flags: InitFlags) -> Result<(), AnyError> {
   create_file(&dir, "main.ts", main_ts)?;
 
   let main_test_ts = include_str!("./templates/main_test.ts")
-    .replace("{CURRENT_STD_URL}", compat::STD_URL_STR);
+    .replace("{CURRENT_STD_URL}", node::STD_URL_STR);
   create_file(&dir, "main_test.ts", &main_test_ts)?;
 
   println!("✅ Project initialized");
